Start of switch to SDL2 (removal of OpenGL).

+++ src/system/wipe.h
@@ -22,4 +22,4 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #define WIPE_SPEED	24
 
-extern void drawFilledRect(int x, int y, int w, int h, float r, float g, float b, float a);
+extern void drawFilledRect(int x, int y, int w, int h, int r, int g, int b, int a);
